Start to handle matplotlib.pyplot for graphs
